move CreateRole to its own file convert handleChange to generic handleInputChange
