+++ src/screens/Roles/CreateRole.tsx
@@ -1,21 +1,17 @@
 import React, { useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
+import { SkillSelection } from './SkillSelection';
 import { createNewRole } from '../../utilities/api/roles';
-import { allSkills, emptyRole } from '../../utilities/data/projectConstants';
+import { emptyRole } from '../../utilities/data/projectConstants';
 import { ProjectInterface } from '../../utilities/types/ProjectInterface';
+import { handleInputChange } from '../../utilities/helpers/formHandlers';
 import './CreateRole.scss';
 
 export const CreateRole = () => {
   const [newRole, setNewRole] = useState(emptyRole);
   const [roleType, setRoleType] = useState<string>();
-  const [selectedSkills, setSelectedSkills] = useState<string[]>([]);
   const { description, maxHeadCount, status, title } = newRole;
   const { id: projectId } = useParams();
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement>) => {
-    const { name, value } = e.target;
-    setNewRole({ ...newRole, [name]: value });
-  };
 
   const handleRoleChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     const { value } = e.target;
@@ -32,46 +28,6 @@ export const CreateRole = () => {
     setNewRole(emptyRole);
   };
 
-  const handleSkillSelection = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    const { value } = e.target;
-    if (!selectedSkills.includes(value)) {
-      let updatedSelectedSkills = [...selectedSkills, value];
-      setSelectedSkills(updatedSelectedSkills);
-    }
-  };
-
-  const SkillSelection = () => {
-    const capitalSelectedSkills: string[] = selectedSkills.map((skill) => skill.toUpperCase());
-
-    return (
-      <div className="skill-selection">
-        {selectedSkills.length > 0 && (
-          <div className="selected-skills--parent">
-            <h4>Selected skills: </h4>
-            <div className="selected-skills--container">
-              {capitalSelectedSkills.map((skill, idx) => (
-                <span className="selected-skills--single" key={`selected-skill-${idx}`}>
-                  {skill}
-                </span>
-              ))}
-            </div>
-          </div>
-        )}
-
-        <select required onChange={handleSkillSelection}>
-          {allSkills.map((skill, idx) => {
-            const upperCaseSkill = skill[0].toUpperCase() + skill.slice(1);
-            return (
-              <option value={skill} key={`option-skill-${idx}`}>
-                {upperCaseSkill}
-              </option>
-            );
-          })}
-        </select>
-      </div>
-    );
-  };
-
   return (
     <div className="new-role--parent">
       <h1>CreateRole</h1>
@@ -82,7 +38,7 @@ export const CreateRole = () => {
             type="text"
             name="title"
             placeholder="My Very Cool Project"
-            onChange={handleChange}
+            onChange={(e) => handleInputChange(e, setNewRole)}
             value={title}
             required
           />
@@ -102,7 +58,7 @@ export const CreateRole = () => {
           <textarea
             name="description"
             placeholder="My Very Cool Project"
-            onChange={handleChange}
+            onChange={(e) => handleInputChange(e, setNewRole)}
             value={description}
             maxLength={300}
             required
@@ -116,7 +72,14 @@ export const CreateRole = () => {
 
         <label>
           Max Headcount
-          <input type="number" name="maxHeadCount" onChange={handleChange} value={maxHeadCount} required min={1} />
+          <input
+            type="number"
+            name="maxHeadCount"
+            onChange={(e) => handleInputChange(e, setNewRole)}
+            value={maxHeadCount}
+            required
+            min={1}
+          />
         </label>
 
         <label>
